fix: refactored namespace
also changed id and updated deps
