+++ src/main/resources/META-INF/services/org.keycloak.storage.UserStorageProviderFactory
@@ -1,1 +1,1 @@
-dk.rmgroup.keycloak.storage.api.ItcnApiUserStorageProviderFactory
+dk.rmgroup.keycloak.storage.api.itcn.ItcnApiUserStorageProviderFactory
